Add DeviceRepository assembly.
The DeviceRepository loads the Model objects.
The Repo uses the XML assembly to get the Model
data, but, can also retrieve data from other
sources. This initial checkin provides a single
Repository class with a static Load() method.
The repository load is invoked by the constructor
for the SystemViewModel class.